Revert MobileTimePicker mousemove test changes
